Add a `Once` that keeps track of cell init status
This field is used to panic on `QrwCell::get` and `QrwCell::get_weak` if
the cell has not been initialized yet (the only case being if the cell
was created with `QrwCell::new`, which is a constant function). This
eliminates the forever-loop behaviour that might be really hard to
debug.